+++ packages/compendium/next/src/components/SearchcraftInit.tsx
@@ -15,7 +15,7 @@ import {
 export const SearchcraftInit = () => {
   useEffect(() => {
     new Searchcraft({
-      index: [process.env.NEXT_PUBLIC_RUNEGARD_INDEX || ''],
+      indexName: process.env.NEXT_PUBLIC_RUNEGARD_INDEX || '',
       readKey: process.env.NEXT_PUBLIC_RUNEGARD_READ_KEY || '',
       endpointURL: process.env.NEXT_PUBLIC_RUNEGARD_ENDPOINT_URL || '',
     });