+++ packages/react-sdk/src/server/index.ts
@@ -3,7 +3,7 @@ import {
   SearchcraftCore,
 } from '@searchcraft/javascript-sdk';
 
-import packageJson from '../../package.json';
+import { name as packageName, version as packageVersion } from '../../package.json';
 
 /**
  * Exports the shared types that all SDKs need
@@ -23,8 +23,8 @@ export class Searchcraft extends SearchcraftCore {
     super(
       config,
       {
-        sdkName: packageJson.name,
-        sdkVersion: packageJson.version,
+        sdkName: packageName,
+        sdkVersion: packageVersion,
       },
       searchcraftId,
     );